requirements.txt
pycryptodome~=3.10.1
python-dateutil~=2.8.1
schedule~=1.0.0